requirements.txt
msal>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
azure-functions>=1.17.0
azure-data-tables>=12.5.0
azure-storage-blob>=12.19.0
pytest>=7.4.0
pytest-cov>=4.1.0
